Rename the shared marg test fixture to cluster with its tests
test/validation.py is the injection-and-noise fixture the marginalization
accuracy and brute-force tests share; the generic name did not say so.
Rename it to test/marg_test_data.py so it sits with test_marg_* and reads
as their support rather than a test of its own.
