+++ test/test_marg_accuracy.py
@@ -42,7 +42,7 @@ import unittest
 
 import numpy
 from utils import simple_exit
-from validation import FLOW, TC, get_seed, make_data
+from marg_test_data import FLOW, TC, get_seed, make_data
 
 from pycbc.distributions import (
     CosAngle,
